backend/requirements.txt
flask
flask-cors
requests
python_dotenv
